refactor: product 도메인별 레이어 추가
