--- src/main/java/kr/hhplus/be/server/product/ProductController.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.product;
-
-public class ProductController {
-}
