+++ src/main/java/kr/hhplus/be/server/product/dto/response/ProductDetailResponse.java
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.product;
+package kr.hhplus.be.server.product.dto.response;
 
-import kr.hhplus.be.server.product.Product;
+import kr.hhplus.be.server.product.domain.Product;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
