+++ src/main/java/kr/hhplus/be/server/product/presentation/ProductController.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.product.presentation;
+
+public class ProductController {
+}
